block: Complete the split betweeen virtio MMIO and block
Creating new modules virtio and mmio that hold the transport trait and
the implementation of the MMIO driver respectively. This will allow the
virtio PCI device driver to be implemented.
